john_deere_demo/pyproject.toml
[project]
name = "demo-agent"
version = "0.1.0"
description = "A modular AI agent system for John Deere equipment sales and support"
readme = "README.md"
authors = [
    { name = "Siddharth Gupta", email = "sid2harthgupta@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "galileo>=1.15.0",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.30",
    "langchain-pinecone>=0.2.11",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.5",
    "pinecone>=7.3.0",
    "pinecone-plugin-interface>=0.0.7",
    "requests>=2.31.0",
    "streamlit>=1.48.1",
]

[project.scripts]
demo-agent = "demo_agent.main:main"

[build-system]
requires = ["uv_build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "pytest>=8.4.1",
    "ruff>=0.12.9",
]
